View cache

Approved-by: Louise Poubel <lupoubel@hotmail.com>

+++ include/ignition/gazebo/Types.hh
@@ -67,7 +67,7 @@ namespace ignition
     /// \brief A unique identifier for a component type. A component type
     /// can be plain data type or something more complex like
     /// ignition::math::Pose3d.
-    using ComponentTypeId = int;
+    using ComponentTypeId = uint64_t;
 
     /// \brief A key that uniquely identifies, at the global scope, a component
     /// instance
